move Autocomplete to sit under deprecated dir
