+++ .storybook/stories.js
@@ -42,8 +42,8 @@ import '../stories/GoogleAddressInput'; // 3.11 GoogleAddressInput
 
 // 4. Selection
 import '../stories/Dropdown'; // 4.1 Dropdown
-import '../stories/StylableAutocomplete'; // 4.1 + StylableAutocomplete
-import '../stories/StylableAutocompleteComposite'; // 4.1 + StylableAutocompleteComposite
+import '../stories/Autocomplete'; // 4.1 + Autocomplete
+import '../stories/AutocompleteComposite'; // 4.1 + AutocompleteComposite
 import '../stories/InputWithOptions'; // 4.1 + InputWithOptions
 import '../stories/IconWithOptions'; // 4.1 + IconWithOptions
 import '../stories/ButtonWithOptions'; // 4.1 + ButtonWithOptions
