+++ stories/Autocomplete/index.js
@@ -8,7 +8,7 @@ import Text from '../../src/Text';
 import {Example} from './ExampleStandard';
 import ExampleStandardRaw from '!raw-loader!./ExampleStandard';
 
-import {Autocomplete} from 'wix-style-react/StylableAutocomplete';
+import Autocomplete from 'wix-style-react/Autocomplete';
 import {generateOptions} from 'wix-ui-core/dist/src/baseComponents/DropdownOption/OptionsExample';
 
 const options = generateOptions((args = {}) => Autocomplete.createDivider(args.value));
@@ -18,7 +18,7 @@ storiesOf('4. Selection', module)
     <TabbedView tabs={['API']}>
       <div>
         <Markdown source={`# \`<Autocomplete/>\``}/>
-        <div style={{background: 'azure', display: 'inline-block'}}><Text>{`import {Autocomplete} from 'wix-style-react/Autocomplete';`}</Text></div>
+        <div style={{background: 'azure', display: 'inline-block'}}><Text>{`import Autocomplete from 'wix-style-react/Autocomplete';`}</Text></div>
         <h2>
           Component documentation and playground is available <a target="_blank" rel="noopener noreferrer" href="https://wix.github.io/wix-ui-backoffice/?selectedKind=Components&selectedStory=Autocomplete">here</a>.
         </h2>
