+++ stories/AutocompleteComposite/ExampleStandard.scss
